requirements.txt
Flask==2.3.3
SQLAlchemy>=2.0.23
Flask-SQLAlchemy>=3.1.1
python-dotenv==1.0.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
alembic>=1.13.0
prometheus-client==0.19.0
flask-caching==2.1.0
redis>=5.0.1  # For production caching